backend/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis[hiredis]>=5.0.1

# Cloud Storage
boto3==1.34.29
pinecone-client==3.0.2

# Security & Utils
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
email-validator==2.1.0.post1

# WebSocket
websockets==12.0

# Existing project dependencies
beautifulsoup4>=4.12.0
chromadb>=0.4.0
moviepy>=1.0.3
numpy>=1.24.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.2
pysrt>=1.1.2
pyyaml>=6.0.1
requests>=2.31.0
tqdm>=4.66.0
nltk>=3.9.0 

openai==1.61.1